rename to vector_triple.h
