+++ src/lib/include/vector_triple.h
